+++ src/eval_framework/run.py
@@ -9,9 +9,10 @@ try:
 except ImportError:
     DeterminedContext = None  # type: ignore
 
+
 from eval_framework.context.local import LocalContext
 from eval_framework.main import main
-from eval_framework.task_loader import load_extra_tasks
+from eval_framework.tasks.task_loader import load_extra_tasks
 from eval_framework.utils import setup_logging
 
 logger = logging.getLogger(__name__)