+++ src/eval_framework/tasks/benchmarks/hellaswag.py
@@ -1,7 +1,7 @@
 import re
 from typing import Any
 
-from eval_framework.metrics.loglikelihood_metrics.accuracy_loglikelihood import (
+from eval_framework.metrics.loglikelihood.accuracy_loglikelihood import (
     AccuracyLoglikelihood,
     AccuracyNormLoglikelihood,
 )