+++ CHANGELOG.md
@@ -4,11 +4,16 @@
 
 ### Added
 
+- 2025-09-02: Add `DOCKER_CODE_EXECUTION` variable to `.env.example`
 - 2025-08-27: Adds support for weights and biases + determined pre-emption
 
 ### Changed
 
+- 2025-09-03: Packages are now released to PyPI
+- 2025-09-03: Import paths in `llm` and `metrics` no longer have a `_llm` and `_metrics` suffix. E.g., `llm/huggingface.py` instead of `llm/huggingface_llm.py`
 - 2025-08-29: commented out the flacky SPHYR test
+- 2025-09-03: Our benchmarks tasks are now registered lazily, which reduces the amount of code that is imported
+              at startup time.
 
 ### Fixed